+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Optimized the dialog sizes for mobile (full screen)
+
 ### Fixed
 
 - Fixed the warning `export was not found` in connection with `GetValuesParams`